Install yarn (to npm global prefix).
This allows the Dashboard (and other projects) to migrate to yarn
which is replacing bower/npm. This also makes using yarn in CI much simpler;
no longer have to npm install yarn first and then attempt to run it
locally etc etc.

Furthermore, gulp-yarn does not work if yarn is installed locally.